+++ src/indices.lisp
@@ -67,7 +67,7 @@
                               :posts (subseq content start end)
                               :title "Recent Posts")))
 
-(defun render-indices (make-symlink-p)
+(defun render-indices ()
   "Render the indices to view content in groups of size N, by month, and by tag."
   (let ((results (by-date (hash-table-values *content*))))
     (dolist (tag (all-tags))
@@ -83,6 +83,4 @@
                                  :prev (and (plusp i) i)
                                  :next (and (< (* (1+ i) 10) (length results))
                                             (+ 2 i)))))))
-  (if make-symlink-p
-    (update-symlink "index.html" "1.html")
-    (run-program "cp 1.html index.html")))
+  (update-symlink "index.html" "1.html"))