+++ NEWS.md
@@ -1,6 +1,7 @@
 ## Changes for 0.9.2 (2013-05-xx):
 
-* A new and improved implementation of tags.
+* A plugin for Github Pages support. (thanks @mrordinaire!)
+* A new and improved implementation of tags. (thanks @woudshoo!)
 * A THEME-DOES-NOT-EXIST error is raised when the theme can't be found.
 
 ## Changes for 0.9.1 (2013-04-10):